pcb: rewire by hand
hah, i actually did it...

The PCB is now hand wired and muuuuuuuch better. This drops the USB pins
on the breakout header because they probably aren't needed and added too
much complexity to the traces.

Also replace the UART pinout with the 5v one on the Atmega side of the
level shifter rather than the 1.8v levels at the phone.
 * This would mean UART is borked unless you attach 5v to the Atmega USB
   port, maybe we want to power the level shifter with the phone vbus
   supply instead ?